ai-service/requirements.txt
# =============================================================================
# CleanGrid AI Service Dependencies
# =============================================================================

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Computer Vision & AI
ultralytics==8.0.206
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.1.0

# HTTP Client for image downloads
httpx==0.25.2
aiofiles==23.2.1

# File handling
python-multipart==0.0.6

# Configuration & Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
numpy==1.24.3
structlog==23.2.0
loguru==0.7.2

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Testing
pytest-httpx==0.26.0
